requirements.txt
ultralytics==8.2.96
supervision==0.1.0
PyYAML==6.0.1
python-box==7.2.0
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
opencv-python==4.9.0.80